pyproject.toml
[tool.poetry]
name = "forced-choice"
version = "0.1.0"
description = ""
authors = ["Loonan Chauvette <loonan.chauvette@gmail.com>"]
readme = "README.md"
packages = [{include = "forced_choice"}]

[tool.poetry.dependencies]
python = "^3.11"
pycaw = "^20230407"
numpy = "^1.24.2"
python-vlc = "^3.0.18121"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
